+++ Reflection/test/utopia/reflection/test/swing/ScrollAreaTest.scala
@@ -51,7 +51,7 @@ object ScrollAreaTest extends App
 
 	// Creates the main stack
 	val stack = Stack.rowWithItems(columns, 16.fixed, 4.fixed)
-	stack.background = Color.yellow.minusHue(Rotation.ofDegrees(33)).darkened(1.2)
+	stack.background = Color.yellow.minusHue(Rotation.ofDegrees(33)).darkened
 
 	val actorHandler = ActorHandler()